requirements.txt
anyio==3.5.0
asgiref==3.5.0
black==22.3.0
certifi==2021.10.8
click==8.0.4
colorama==0.4.4
colorlog==6.6.0
colormap==1.0.4
cycler==0.11.0
easydev==0.12.0
fastapi==0.75.0
fonttools==4.31.1
h11==0.13.0
idna==3.3
imageio==2.16.1
joblib==1.1.0
kiwisolver==1.4.0
mypy-extensions==0.4.3
networkx==2.7.1
numpy==1.22.3
packaging==21.3
pathspec==0.9.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
ptyprocess==0.7.0
pydantic==1.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-multipart==0.0.5
PyWavelets==1.3.0
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
threadpoolctl==3.1.0
tifffile==2022.3.16
tomli==2.0.1
typing_extensions==4.1.1
uvicorn==0.17.6
wincertstore==0.2
